Add single-item removal and dates for game releases

+++ main.py
@@ -4,12 +4,13 @@ db = database_handler.Collection_Database()
 
 choice = 0
 
-while choice != 4:
+while choice != 5:
 
     print("1. Add item to database")
     print("2. Wipe database")
-    print("3. View video_game table")
-    print("4. Quit")
+    print("3. View database")
+    print("4. Remove items from database")
+    print("5. Quit")
 
     choice = int(input("Select an option: "))
 
@@ -30,8 +31,29 @@ while choice != 4:
             title = input("What book?: ")
             db.register_item("books", title)
     elif choice == 2:
-        db.wipe_database()
+        choice = input("Are you sure? y/n: ")
+        if choice == 'y':
+            db.wipe_database()
     elif choice == 3:
         db.view_video_games()
         db.view_tabletop_rpgs()
         db.view_books()
+    elif choice == 4:
+        print("Removing items from database")
+        print("1. Video game")
+        print("2. Tabletop RPG")
+        print("3. Book")
+        print("4. Back out")
+        sub_choice = int(input("Select an option: "))
+        if sub_choice == 1:
+            db.view_video_games()
+            ids = input("List video games to remove by ID, separated by commas: ")
+            db.deregister_item("video_games", ids)
+        elif sub_choice == 2:
+            db.view_tabletop_rpgs()
+            ids = input("List tabletop RPGs to remove by ID, separated by commas:")
+            db.deregister_item("tabletop_rpgs", ids)
+        elif sub_choice == 3:
+            db.view_books()
+            ids = input("List books to remove by ID, separated by commas:")
+            db.deregister_item("books", ids)
